added missing sqleet.h renamed config.c to sqleet-config.h
