+++ sqleet.h
@@ -1,0 +1,2 @@
+#include "sqleet-config.h"
+#include "sqlite3.h"
